Separate geometric predicate manager term
