+++ isaaclab_arena/tasks/pick_and_place_task.py
@@ -23,12 +23,9 @@ from isaaclab_arena.metrics.object_moved import ObjectMovedRateMetric
 from isaaclab_arena.metrics.success_rate import SuccessRateMetric
 from isaaclab_arena.progress_tracking.progress_objective import ProgressObjective
 from isaaclab_arena.tasks.common.mimic_default_params import MIMIC_DATAGEN_CONFIG_DEFAULTS
+from isaaclab_arena.tasks.predicates.geometric_object_on_destination_term import GeometricObjectOnDestinationTerm
 from isaaclab_arena.tasks.predicates.object_settling import objects_settled
-from isaaclab_arena.tasks.predicates.spatial import (
-    GeometricObjectOnDestinationTerm,
-    object_is_above_height,
-    object_on_destination,
-)
+from isaaclab_arena.tasks.predicates.spatial import object_is_above_height, object_on_destination
 from isaaclab_arena.tasks.task_base import TaskBase
 from isaaclab_arena.tasks.task_transition import Relocate, TaskTransition
 from isaaclab_arena.utils.cameras import get_viewer_cfg_look_at_object
@@ -132,6 +129,7 @@ class PickAndPlaceTask(TaskBase):
         return self.termination_cfg
 
     def make_termination_cfg(self):
+        # The class-backed term caches spawned object and destination bounds once for the live environment.
         success = TerminationTermCfg(
             func=GeometricObjectOnDestinationTerm,
             params={
